Refactor: python, cts: Move ComponentFail into the _cts module.
No code changes.
