+++ python/pacemaker/_cts/tests/__init__.py
@@ -5,6 +5,7 @@ Test classes for the `pacemaker._cts` package.
 __copyright__ = "Copyright 2023 the Pacemaker project contributors"
 __license__ = "GNU Lesser General Public License version 2.1 or later (LGPLv2.1+)"
 
+from pacemaker._cts.tests.componentfail import ComponentFail
 from pacemaker._cts.tests.ctstest import CTSTest
 from pacemaker._cts.tests.fliptest import FlipTest
 from pacemaker._cts.tests.maintenancemode import MaintenanceMode
